skills/rhanbourinajd/ai-video-gen/requirements.txt
openai>=1.0.0
replicate>=0.20.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
